% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cTWAS.R
\name{estimate_priors_EM}
\alias{estimate_priors_EM}
\title{cTWAS EM algorithm to optimize priors}
\usage{
estimate_priors_EM(zscores, R, L = 1, n_iter = 30)
}
\arguments{
\item{zscores}{A numeric vector of z-score association statistics for the
variables on the trait of interest.}

\item{R}{A numeric matrix or sparse Matrix object of correlations between
variables}

\item{L}{The maximum number of effects (credible sets) Susie-RSS is allowed
to fit.}

\item{n_iter}{Number of EM iterations to run.}
}
\value{
A list object with the following elements:

\describe{
\item{pi}{A list containing the updated cTWAS prior inclusion probability
(prior probability of effect on the trait) for each class of variables}

\item{sigma2}{A list containing the updated cTWAS prior effect variances,
conditioned on the variable having an effect, for each class of variables}

\item{n_vars}{A list object giving the number of variables in each class,
e.g. ('SNPs' = 10000, 'exposures' = 30)}
}
}
\description{
Implements the cTWAS Expectation-Maximization (EM) algorithm for fitting
prior probabilities and variances for variables of each "class" (SNPs and
exposures/factors). Each iteration runs Susie-RSS with the current priors,
then updates the priors given the Susie-RSS results.
}
