% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{famr}
\alias{famr}
\title{Perform a family of multiple regression analyses on a dataset}
\usage{
famr(data, y, x, models = NULL)
}
\arguments{
\item{data}{- matrix or data frame}

\item{y}{- character, name of column in data containing modeled variable;
alternatively, a list with one function operating on data.}

\item{x}{- character, name of column in data containing primary variable;
alternatively, a list with one function operating on data.}

\item{models}{- description of secondary variables to include in
multiple regressions.
When NULL, all numeric/factor columns in data are used.
Otherwise, models must be a named list, see output of famrModels() for
specific format.}
}
\description{
Perform a family of multiple regression analyses on a dataset
}
