% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{famrModels}
\alias{famrModels}
\title{Prepare an object with a definition of models}
\usage{
famrModels(data, avoid = NULL)
}
\arguments{
\item{data}{- data frame with column names}

\item{avoid}{- vector with columns to avoid}
}
\value{
a nested list, suitable for passing as
        'other' into function famr()
}
\description{
Prepare an object with a definition of models
}
