% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{famrPrep}
\alias{famrPrep}
\title{Prepare a dataset for mr analysis.}
\usage{
famrPrep(data, y, x, models = NULL)
}
\arguments{
\item{data}{- see help for famr()}

\item{y}{- see help for famr()}

\item{x}{- see help for famr()}

\item{models}{- see help for famr()

(this is an internal helper function)}
}
\description{
The main functionality here is to allow applying functions
on the input data.
}
