% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_X}
\alias{impute_X}
\title{Impute values under the limit of detection (LOD) in mixtures}
\usage{
impute_X(prm, X, Ilod, loda, lodb, Xmean)
}
\arguments{
\item{Ilod}{is (n x p) indicator matrix of observations below lod}

\item{loda}{is a vector of p lower limit LOD for variables in X}

\item{lodb}{is a vector of p upper limit LOD for variables in X}
}
\description{
Impute values under the limit of detection (LOD) in mixtures
}
