% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ld_prune_famr}
\alias{ld_prune_famr}
\title{LD clump SNPs in a locus}
\usage{
ld_prune_famr(sumstats, ld, prune_thresh)
}
\arguments{
\item{sumstats}{A summary statistics list object, containing the
following fields (and possibly others):

\describe{
\item{betas}{A matrix of estimated effect sizes of SNPs on exposures, with
rows corresponding to SNPs and columns corresponding to exposures.}

\item{stderrs}{A matrix of standard errors for the betas.}

\item{Z}{A matrix of z-scores for the exposures, equal to betas divided by
stderrs.}
}}

\item{ld}{A numeric matrix of LD values between SNPs in sumstats}

\item{prune_thresh}{A numeric value between 0 and 1 indicating the LD
clumping threshold.}
}
\value{
A vector of SNP indices marked to be retained after LD clumping.
}
\description{
A simple helper function to LD clump SNPs in a locus. Given a clumping
threshold and a summary statistics object and LD matrix, returns a list of
SNP indices to keep after clumping. SNPs are sorted in descending order by
highest Z-score with any exposure, and SNPs are marked for removal when they
have high enough LD with a SNP with a stronger peak Z-score.
The kept indices are usually passed to subset_sumstats to actually
remove the SNPs.
}
