% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{learn_lambda}
\alias{learn_lambda}
\title{Learn regularization parameters for Susie-RSS weight learning}
\usage{
learn_lambda(sumstats, ld, oracle_mode = F)
}
\arguments{
\item{sumstats}{A list object, usually read by gather_sumstats. The list has
three matrices with SNPs as rows and exposure traits as columns. These
are "betas" (effect size estimates of SNPs on exposures), "stderrs"
(standard errors of the betas), and "Z" (the z-scores, i.e. betas/stderrs).
See gather_sumstats for more details.}

\item{ld}{numeric LD matrix corresponding to the SNPs in sumstats}

\item{oracle_mode}{DEPRECATED. used only for internal testing. in this case,
it automatically returns 0 for the lambda parameters, i.e. no regularization.}
}
\value{
a numerical vector of learned lambda parameters for each trait for
this locus. each is between 0 and 1.
}
\description{
This function learns a "lambda" parameter to regularize the LD matrix of SNPs
in case of a mismatch in populations between an LD reference panel and the
population the summary statistics were gathered in. This improves the
stability of weights learned by Susie-RSS in learn_weights. The lambdas
will be between 0 and 1 and there is one learned for each trait in each
independent locus. lambda=0 is no regularization, lambda=1 sets R to the
identity matrix.
}
