% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_outcome}
\alias{merge_outcome}
\title{Merge outcome and exposure summary statistics}
\usage{
merge_outcome(sumstats, ld, y_gwas)
}
\arguments{
\item{sumstats}{A list object with the following elements, i.e. summary
statistics for exposures:

\describe{
\item{betas}{A matrix of estimated effect sizes of SNPs on exposures, with
rows corresponding to SNPs and columns corresponding to exposures.}

\item{weights}{Currently this is the same as betas; later used to store
learned weights of SNPs on exposures and factors.}

\item{stderrs}{A matrix of standard errors for the betas.}

\item{Z}{A matrix of z-scores for the exposures, equal to betas divided by
stderrs.}

\item{pos}{A vector of SNP names.}
}}

\item{ld}{A numeric LD matrix between the SNPs in sumstats.}

\item{y_gwas}{A list object with the following elements, i.e. summary
statistics for the outcome:

\describe{
\item{names_y}{A vector of SNP IDs (names)}

\item{betas_y}{A vector of estimated effect sizes of SNPs on the outcome.}

\item{stderrs_y}{A vector of standard errors for betas_y.}

\item{zscores_y}{A vector of z-scores of SNPs on the outcome, equal to
betas_y divided by stderrs_y.}
}}
}
\value{
A list object with the following elements:

\describe{
\item{sumstats}{A list with the fields of both sumstats and y_gwas (except
names_y, which is merged into pos), containing the merged summary statistics}

\item{ld}{A numeric LD matrix for the SNPs in the new sumstats object;
the same as the input LD matrix except without the SNPs that are unmeasured
in y_gwas.}
}
}
\description{
Given exposure summary statistics and outcome summary statistics,
intersects the SNPs, then merges them into a single object. Used if
read_y_gwas is run, i.e. if outcome summary statistics are not already
included in the exposure summary statistics files.
}
\keyword{internal}
