% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_sumstats}
\alias{merge_sumstats}
\title{Merge sumstats for a locus into overall sumstats}
\usage{
merge_sumstats(all_sumstats, locus_sumstats, by_row = T)
}
\arguments{
\item{all_sumstats}{A summary statistics list object, containing some or all
of the following fields (only betas, stderrs, weights, Z required):

\describe{
\item{betas}{A matrix of estimated effect sizes of SNPs on exposures, with
rows corresponding to SNPs and columns corresponding to exposures.}

\item{weights}{A matrix of weights learned for each SNP on each exposure,
with rows corresponding to SNPs and columns corresponding to exposures.}

\item{stderrs}{A matrix of standard errors for the betas.}

\item{Z}{A matrix of z-scores for the exposures, equal to betas divided by
stderrs.}

\item{betas_y}{A vector of estimated effect sizes of SNPs on the outcome.}

\item{stderrs_y}{A vector of standard errors for betas_y.}

\item{zscores_y}{A vector of z-scores of SNPs on the outcome, equal to
betas_y divided by stderrs_y.}

\item{pos}{A vector of SNP names.}

\item{locus_idx}{A vector matching SNPs to which locus they come from,
indexed in the order the loci are read from in_dir.}

\item{lambda}{A matrix of lambda values learned by learn_lambda in famr.R for
each exposure in each locus, where each row of the matrix corresponds to a
locus and the columns are the exposures.}
}}

\item{locus_sumstats}{Like all_sumstats, but for the current locus only.}

\item{by_row}{Boolean indicating whether the sumstats should be merged by row
or by column. The latter is used when merging exposure and factor summary
statistics rather than merging different loci.}
}
\value{
A list object in the same format as all_sumstats, with the merged
summary statistics from all_sumstats and locus_sumstats.
}
\description{
This function is meant to be called from several functions in famr.R that
loop through loci. Given an object containing summary statistics merged from
all loci so far (all_sumstats) and summary statistics from a new locus,
merges these two and returns a new summary statistics object.
Also used to merge exposure and factor summary statistics.
}
\keyword{internal}
