% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mrdata2}
\alias{mrdata2}
\title{Toy dataset for multiple regression analysis}
\format{A matrix with eight columns: \code{w1}-\code{w6}, \code{x}, \code{y}}
\usage{
data(mrdata2)
}
\description{
System wherein y is related to x and w2; other variables are noise
}
\keyword{datasets}
