% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mrdata3}
\alias{mrdata3}
\title{Toy dataset for multiple regression analysis}
\format{A matrix with eight columns: \code{w1}-\code{w6}, \code{x}, \code{y}}
\usage{
data(mrdata3)
}
\description{
System wherein x and y both depend on w1; other variables are noise
}
\keyword{datasets}
