% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mrdata5}
\alias{mrdata5}
\title{Toy dataset for multiple regression analysis}
\format{A matrix with eight columns: \code{w1}-\code{w6}, \code{x}, \code{y}}
\usage{
data(mrdata5)
}
\description{
System wherein x and y both depend on a categorical variable;
other variables have correlations among them
}
\keyword{datasets}
