% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.famr}
\alias{plot.famr}
\title{Produce a visual representation of a multiple regression}
\usage{
\method{plot}{famr}(x, type = c("scatter", "pvalues"), label.pdiff = 1,
  Rcss = "famr", Rcssclass = "famr", ...)
}
\arguments{
\item{x}{- a famr object}

\item{type}{- character, instruction as to what kind of chart to draw}

\item{label.pdiff}{- threshold above which to display text labels
on the pvalues chart}

\item{Rcss}{- style object used by Rcssplot}

\item{Rcssclass}{- style class used by Rcssplot}

\item{...}{- additional graphical arguments used in base plot,
e.g. xlim and ylim values}
}
\description{
Produce a visual representation of a multiple regression
}
