% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFamrPvalues}
\alias{plotFamrPvalues}
\title{Plot a x-y scatter chart with model p-values}
\usage{
plotFamrPvalues(x, label.pdiff = 1.3, ...)
}
\arguments{
\item{x}{- famr object}

\item{label.pdiff}{- numerical, determines how far
away a dot has to be from base model to include a label}

\item{...}{- additional parameters passed on to graphics::plot()}
}
\description{
This is an internal function invoked through plot()
}
