% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFamrScatter}
\alias{plotFamrScatter}
\title{Plot a x-y scatter chart with linear models}
\usage{
plotFamrScatter(x, ...)
}
\arguments{
\item{x}{- famr object}

\item{...}{- additional parameters passed on to graphics::plot()}
}
\description{
This is an internal function invoked through plot()
}
