% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{print.famr}
\alias{print.famr}
\title{Display basic information about the model family}
\usage{
\method{print}{famr}(x, ...)
}
\arguments{
\item{x}{- famr object}

\item{...}{- additional paramters (not used)}
}
\description{
Display basic information about the model family
}
