% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prune_and_merge}
\alias{prune_and_merge}
\title{Wrapper function to clump and merge SNPs}
\usage{
prune_and_merge(
  dat,
  sumstats,
  ld,
  prune_thresh,
  fa_prune_thresh,
  f_idx,
  oracle_mode = F
)
}
\arguments{
\item{dat}{A list object formed from merging all previous loci,
with the following elements:

\describe{
\item{betas}{A matrix of estimated effect sizes of SNPs on exposures, with
rows corresponding to SNPs and columns corresponding to exposures.}

\item{stderrs}{A matrix of standard errors for the betas.}

\item{Z}{A matrix of z-scores for the exposures, equal to betas divided by
stderrs.}

\item{betas_y}{A vector of estimated effect sizes of SNPs on the outcome.}

\item{stderrs_y}{A vector of standard errors for betas_y.}

\item{zscores_y}{A vector of z-scores of SNPs on the outcome, equal to
betas_y divided by stderrs_y.}

\item{pos}{A vector of SNP names.}

\item{locus_idx}{A vector matching SNPs to which locus they come from,
indexed in the order the loci are read from in_dir.}

\item{lambda}{A matrix of lambda values learned by learn_lambda for each
exposure in each locus, where each row of the matrix corresponds to a locus
and the columns are the exposures.}

\item{ss_for_fa}{A list object with the "betas", "stderrs", "Z", "weights",
and "pos" fields, but only for the SNPs to be input to factor analysis.}
}}

\item{sumstats}{A list object in the same format as dat, but corresponding
to the current locus}

\item{ld}{A numertic LD matrix between the SNPs in sumstats (current locus)}

\item{prune_thresh}{A numeric value between 0 and 1 indicating the LD
clumping threshold to be used by ld_prune_famr}

\item{fa_prune_thresh}{A numeric value between 0 and 1 indicating the LD
clumping threshold to be used to generate a set of SNPs for input to factor
analysis (FA)}

\item{f_idx}{The current file index, used to indicate which locus the SNPs in
sumstats come from}

\item{oracle_mode}{DEPRECATED. Used for internal testing only.}
}
\value{
A list object in the same format as dat, containing the SNPs in both
dat and those from sumstats that were not removed during LD clumping.
}
\description{
A simple wrapper function that performs LD clumping for SNPs in a locus,
then subsets those SNPs, then merges the SNPs for this locus into an object
collecting summary statistics over all loci. Used by gather_sumstats to read
individual locus files into a single object in memory.
}
\keyword{internal}
