% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{read_y_gwas}
\alias{read_y_gwas}
\title{Read in outcome trait summary statistics}
\usage{
read_y_gwas(y_gwas_file, idcol, betacol, secol, header = T)
}
\arguments{
\item{y_gwas_file}{File path to the file with the outcome
summary statistics. Can either be a TSV file with columns for the SNP name,
effect size estimates, and standard errors (see idcol, betacol, secol),
or a VCF file in the MRC IEU GWAS format, described here:
https://github.com/MRCIEU/gwas-vcf-specification
See also read_vcf in utils.R.}

\item{idcol}{Column of y_gwas_file with the SNP IDs (names). Not used if a
VCF is provided.}

\item{betacol}{Column of y_gwas_file with the SNP effect sizes (betas).
Not used if a VCF is provided.}

\item{secol}{Column of y_gwas_file with the SNP effect standard errors.
Not used if a VCF is provided.}

\item{header}{Boolean indicating whether the y_gwas_file has a header line,
e.g. with column names. Default=TRUE. Not used if a VCF is provided.}
}
\value{
Returns a list with four items: names_y, betas_y, stderrs_y, and
zscores_y, giving the IDs, estimated effect sizes, standard errors, and
z-scores of the SNPs on the outcome trait, respectively.
}
\description{
Optional helper function to read in summary statistics for the outcome trait,
used if they are not already provided in the summary statistics files read
by in_dir (see famr_susie function documentation).
}
