% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_gfa_full}
\alias{run_gfa_full}
\title{Run GFA for factor analysis}
\usage{
run_gfa_full(x_betas, x_stderrs, N)
}
\arguments{
\item{x_betas}{Numeric matrix of marginal effect size estimates ("betas") of
SNPs on each exposure, with rows corresponding to SNPs and columns
corresponding to exposures.}

\item{x_stderrs}{Numeric matrix of standard errors of the x_betas.}

\item{N}{DEPRECATED. Sample size used to generate the exposures.}
}
\value{
A GFA object; see GFA for more details. For FAMR, the most relevant
fields are L_hat and F_hat, which contain the loadings and factors inferred
by GFA. L_hat will be of dimension nSNPs-by-nFactors, and t(F_hat) will be of
dimension nFactors-by-nExposures.
}
\description{
Wrapper function to run GFA to infer factors for FAMR, while catching errors
that happen when no factors are inferred and removing factors strongly
associated with only one factor.
}
