% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{run_modified_ctwas}
\alias{run_modified_ctwas}
\title{Run cTWAS-style regression framework}
\usage{
run_modified_ctwas(dat, L, n_iter, n_expo, n_samp, annih)
}
\arguments{
\item{dat}{A list object, usually from learn_wts_precomp_merge, with the
following elements:

\describe{
\item{sumstats}{A list object containing summary statistics for all exposures
and factors; see the all_sumstats parameter of learn_wts_precomp_merge
for the format.}

\item{ld}{A sparse Matrix object with the LD for the SNPs in the sumstats
object.}

\item{numerator}{A numeric vector representing the numerator of the Z-scores
for each exposure and factor.}

\item{wRw}{A numeric matrix of associations between traits (exposures and
factors), used for computing both Z and R}

\item{Rgx}{A numeric matrix of inferred correlations between SNPs and
traits (exposures and outcomes), with one row per SNP and one column per
exposure/factor.}
}}

\item{L}{a numeric value representing the maximum number of effects (more
precisely, credible sets) Susie-RSS can model on the outcome}

\item{n_iter}{number of cTWAS expectation-maximization (EM) iterations to
run for fitting prior parameters; usually 10-30 is good.}

\item{n_expo}{number of exposures being analyzed}

\item{n_samp}{number of samples used to generate exposure summary statistics}

\item{annih}{boolean indicating whether the annihilate flag was used
(whether factors were regressed out of exposures and outcome). See
annihilate_factors for more details.}
}
\value{
A list with some or all of the following elements:

\describe{
\item{pips}{A list of the learned PIPs for each exposure, factor, and SNP.}

\item{zscores}{A list of the marginal association z-scores learned for each
exposure, factor, and SNP with the outcome.}

\item{R}{A matrix of the estimated correlations between each pair of
variables (exposures, factors, and SNPs).}

\item{susieres}{The full susie_rss results object. See the susieR
documentation for more details.}

\item{priors}{A list containing the learned prior effect probabilities and
prior effect variances learned for each class of variables (i.e. SNPs and
exposures/factors).}

\item{posterior_mean}{A list containing the posterior means learned for
each variable by susie_rss.}

\item{posterior_var}{A list containing the posterior variances learned for
each variable by susie_rss.}

\item{factor_corrs}{A matrix of correlations between the learned factor
summary statistics and those of the exposures. Deprecated in favor of the
R matrix.}
}
}
\description{
Given the output from learn_wts_precomp_merge, computes the final Z-score
vector and R matrix, uses a cTWAS-style framework to infer priors for each
"class" of variables (SNPs and exposures/factors), and then generates the
final PIPs for each variable using Susie-RSS. Also runs a separate
factor-only regression to isolate factor effects from exposure effects.
}
