% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_sumstats}
\alias{subset_sumstats}
\title{Subset summary statistics objects}
\usage{
subset_sumstats(ss, idx)
}
\arguments{
\item{ss}{A summary statistics list object, containing some or all of the
following fields (only betas, stderrs, weights, Z required):

\describe{
\item{betas}{A matrix of estimated effect sizes of SNPs on exposures, with
rows corresponding to SNPs and columns corresponding to exposures.}

\item{stderrs}{A matrix of standard errors for the betas.}

\item{Z}{A matrix of z-scores for the exposures, equal to betas divided by
stderrs.}

\item{betas_y}{A vector of estimated effect sizes of SNPs on the outcome.}

\item{stderrs_y}{A vector of standard errors for betas_y.}

\item{zscores_y}{A vector of z-scores of SNPs on the outcome, equal to
betas_y divided by stderrs_y.}

\item{pos}{A vector of SNP names.}

\item{locus_idx}{A vector matching SNPs to which locus they come from,
indexed in the order the loci are read from in_dir.}

\item{annih_y}{A vector containing the portion of betas_y projected out
by factors; see annihilate_factors in famr.R for more details}

\item{orig_betas_y}{A vector containing the original betas_y if the latter
was modified by annihilating factors; see annihilate_factors in famr.R
for more details}

\item{orig_zscores_y}{Similar to orig_betas_y but for zscores_y}
}}

\item{idx}{A vector of SNP indices to retain from ss}
}
\value{
A modified version of ss with only the SNPs specified by idx kept.
}
\description{
Subset a summary statistics object ("ss") according to the indices provided
by idx, i.e. retain only the SNPs at the positions given by idx.
}
\keyword{internal}
