% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famr.R
\name{summary.famr}
\alias{summary.famr}
\title{Display a matrix of information summarizing the models}
\usage{
\method{summary}{famr}(object, ...)
}
\arguments{
\item{object}{- famr object}

\item{...}{- additional parameters (not used)}
}
\description{
Display a matrix of information summarizing the models
}
