% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cTWAS.R
\name{update_pi_sigma2}
\alias{update_pi_sigma2}
\title{Update cTWAS priors}
\usage{
update_pi_sigma2(alpha, tau2)
}
\arguments{
\item{alpha}{A list object with elements for each class of variables,
containing a numeric vector of PIPs for the variables of that class.
Usually generated by update_alpha_tau2.}

\item{tau2}{A list object with elements for each class of variables,
containing a numeric vector of posterior second moments (conditioned on the
variable having an effect) for the variables of that class.
Usually generated by update_alpha_tau2.}
}
\value{
A list object with the following elements:

\describe{
\item{pi}{A list containing the updated cTWAS prior inclusion probability
(prior probability of effect on the trait) for each class of variables}

\item{sigma2}{A list containing the updated cTWAS prior effect variances,
conditioned on the variable having an effect, for each class of variables}
}
}
\description{
Given updated posterior parameters learned from running Susie-RSS given the
previous cTWAS priors, update the cTWAS prior inclusion probability and
prior variance for each class of variables (i.e. SNPs and exposures/factors).
Part of the cTWAS EM algorithm.
}
