% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_low_instrument}
\alias{warn_low_instrument}
\title{Warn user about traits with low number of instruments}
\usage{
warn_low_instrument(weights, trait_names, min_instruments = 10)
}
\arguments{
\item{weights}{A matrix of learned instrument-exposure/factor weights}

\item{trait_names}{A vector of the names of the traits}

\item{min_instruments}{Minimum number of instruments; traits with fewer
instruments will trigger a warning}
}
\value{
A vector of trait names with fewer than min_instruments instruments
with non-zero weights
}
\description{
Traits with few instruments with nonzero weights may be poorly predicted
and thus provide somewhat less reliable results in the MR analysis. This
helper functions warns users when this is the case.
}
