% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_famr_res}
\alias{write_famr_res}
\title{Write FAMR-Susie results to a file}
\usage{
write_famr_res(out, famr_res, K, fa_method, for_sim = F)
}
\arguments{
\item{out}{Base name for output files}

\item{famr_res}{An FAMR-Susie results object; see the famr_susie method in
famr.R for more details.}

\item{K}{The number of exposures in the analysis.}

\item{fa_method}{Which factor analysis method was used to infer factors
for FAMR-Susie.}

\item{for_sim}{A boolen indicating whether to generate an additional text
file output for use with internal simulations.}
}
\value{
nothing
}
\description{
Writes FAMR-Susie results to a file, and optionally a simple results summary
to an additional text file. Mainly intended to be used internally for
simulations.
}
